+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+## [v0.303.0] - 2025-11-17
+
 ### Added
 
 - `it-sdi-v1`: Add validation for IBANs.